Remove isValid check and pass modified hex
I think this is the correct behaviour we want. If hex is valid, it changes the
state of the colorpickr. Otherwise the last valid color is shown and the
hex input continues to display what the user typed - allowing them to
complete the input with a valid value.

Closes #100 where typing a fourth value in the hex input after a valid
one would reset the value back to the valid one.